% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serenity.R
\name{loadr}
\alias{loadr}
\title{Load an rda or rds file and add it to the serenity data list (r_data) if available}
\usage{
loadr(fn, objname = "")
}
\arguments{
\item{fn}{File name and path as a string. Extension must be either rda or rds}

\item{objname}{Name to use for the data frame. Defaults to the file name}
}
\value{
Data frame in r_data or in the calling enviroment
}
\description{
Load an rda or rds file and add it to the serenity data list (r_data) if available
}
