% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{serenity.data-deprecated}
\alias{serenity.data-deprecated}
\alias{mutate_each}
\alias{dfprint}
\alias{nrprint}
\alias{serenity.data-deprecated-package}
\title{Deprecated function(s) in the serenity.data package}
\usage{
mutate_each(...)
}
\arguments{
\item{...}{Parameters to be passed to the updated functions}
}
\description{
These functions are provided for compatibility with previous versions of
serenity. They will eventually be  removed.
}
\section{Details}{

\tabular{rl}{
  \code{mutate_each} is now a synonym for \code{\link{mutate_ext}}, \code{\link{mutate_at}}, or \code{\link{mutate_all}}\cr
  \code{dfprint} is now a synonym for \code{\link{formatdf}}\cr
  \code{nrprint} is now a synonym for \code{\link{formatnr}}\cr
  \code{varp_rm} is now a synonym for \code{\link{varpop}}\cr
  \code{sdp_rm} is now a synonym for \code{\link{sdpop}}\cr
}
}

