% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/serenity.R
\docType{package}
\name{serenity.data}
\alias{serenity.data}
\alias{serenity.data-package}
\alias{serenity.data}
\title{serenity.data}
\usage{
serenity.data()
}
\description{
serenity.data

Launch Serenity in the default browser
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
